Update package.json version
In order to sync this up for people using npm, you'll also want to move the v0.1.7 tag forward to this commit.

You might also consider adding a "v" in front of your tags (v0.1.7) if you want to support npm better.  But I understand you might mostly be interested in bower. :)